requirements_pi.txt
# Raspberry Pi Optimized Dependencies

# Core (lightweight)
requests>=2.31.0
python-dotenv>=1.0.0

# Local LLM (Ollama)
ollama>=0.1.0

# Voice & Audio (Pi optimized)
pyaudio>=0.2.14
# Note: On Pi, may need: sudo apt-get install portaudio19-dev python3-pyaudio
webrtcvad>=2.0.10
pocketsphinx>=5.0.0
# Note: On Pi: sudo apt-get install pocketsphinx pocketsphinx-en-us

# Offline STT
vosk>=0.3.45
# Download model: https://alphacephei.com/vosk/models

# Offline TTS
piper-tts>=1.0.0
# Alternative: pyttsx3 (uses system TTS)
pyttsx3>=2.90

# Utilities
numpy>=1.24.0  # Required by audio libs

# Optional: For cloud fallback
openai>=1.12.0
anthropic>=0.18.0

# Cybersecurity testing (authorized use only)
pynput>=1.7.6
cryptography>=41.0.0
requests>=2.31.0

# System audio (Pi)
# sudo apt-get install alsa-utils
